+++ test_conformance/images/kernel_read_write/test_read_1D.cpp
@@ -522,9 +522,12 @@ int test_read_image_1D( cl_context context, cl_command_queue queue, cl_kernel ke
                                                                             xOffsetValues[ j ], 0.0f, 0.0f, norm_offset_x, 0.0f, 0.0f,
                                                                             imageSampler, expected, 0, &containsDenormals, lod );
 
-                            float err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                            float err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                            float err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                            float err1 = fabs(sRGBmap(resultPtr[0])
+                                              - sRGBmap(expected[0]));
+                            float err2 = fabs(sRGBmap(resultPtr[1])
+                                              - sRGBmap(expected[1]));
+                            float err3 = fabs(sRGBmap(resultPtr[2])
+                                              - sRGBmap(expected[2]));
                             float err4 = fabsf( resultPtr[3] - expected[3] );
 
                             float maxErr = 0.5;
@@ -544,9 +547,12 @@ int test_read_image_1D( cl_context context, cl_command_queue queue, cl_kernel ke
                                                                                  xOffsetValues[ j ], 0.0f, 0.0f, norm_offset_x, 0.0f, 0.0f,
                                                                                  imageSampler, expected, 0, NULL, lod );
 
-                                    err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                    err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                    err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                    err1 = fabs(sRGBmap(resultPtr[0])
+                                                - sRGBmap(expected[0]));
+                                    err2 = fabs(sRGBmap(resultPtr[1])
+                                                - sRGBmap(expected[1]));
+                                    err3 = fabs(sRGBmap(resultPtr[2])
+                                                - sRGBmap(expected[2]));
                                     err4 = fabsf( resultPtr[3] - expected[3] );
                                 }
                             }
@@ -576,9 +582,12 @@ int test_read_image_1D( cl_context context, cl_command_queue queue, cl_kernel ke
                                                                                         xOffsetValues[ j ], 0.0f, 0.0f, norm_offset_x, 0.0f, 0.0f,
                                                                                         imageSampler, expected, 0, &containsDenormals, lod );
 
-                                float err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                float err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                float err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                float err1 = fabs(sRGBmap(resultPtr[0])
+                                                  - sRGBmap(expected[0]));
+                                float err2 = fabs(sRGBmap(resultPtr[1])
+                                                  - sRGBmap(expected[1]));
+                                float err3 = fabs(sRGBmap(resultPtr[2])
+                                                  - sRGBmap(expected[2]));
                                 float err4 = fabsf( resultPtr[3] - expected[3] );
 
                                 float maxErr = 0.6;
@@ -597,9 +606,12 @@ int test_read_image_1D( cl_context context, cl_command_queue queue, cl_kernel ke
                                                                                      xOffsetValues[ j ], 0.0f, 0.0f, norm_offset_x, 0.0f, 0.0f,
                                                                                      imageSampler, expected, 0, NULL, lod );
 
-                                        err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                        err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                        err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                        err1 = fabs(sRGBmap(resultPtr[0])
+                                                    - sRGBmap(expected[0]));
+                                        err2 = fabs(sRGBmap(resultPtr[1])
+                                                    - sRGBmap(expected[1]));
+                                        err3 = fabs(sRGBmap(resultPtr[2])
+                                                    - sRGBmap(expected[2]));
                                         err4 = fabsf( resultPtr[3] - expected[3] );
                                     }
                                 }
