+++ test_conformance/images/kernel_read_write/test_read_1D_array.cpp
@@ -618,9 +618,12 @@ int test_read_image_1D_array( cl_context context, cl_command_queue queue, cl_ker
                                 xOffsetValues[ j ], yOffsetValues[ j ], 0.f, norm_offset_x, norm_offset_y, 0.0f,
                                 imageSampler, expected, 0, &containsDenormals, lod );
 
-                            float err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                            float err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                            float err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                            float err1 = fabs(sRGBmap(resultPtr[0])
+                                              - sRGBmap(expected[0]));
+                            float err2 = fabs(sRGBmap(resultPtr[1])
+                                              - sRGBmap(expected[1]));
+                            float err3 = fabs(sRGBmap(resultPtr[2])
+                                              - sRGBmap(expected[2]));
                             float err4 = fabsf( resultPtr[3] - expected[3] );
                             float maxErr = 0.5;
 
@@ -639,9 +642,12 @@ int test_read_image_1D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                xOffsetValues[ j ], yOffsetValues[ j ], 0.f, norm_offset_x, norm_offset_y, 0.0f,
                                                                                imageSampler, expected, 0, NULL, lod );
 
-                                    err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                    err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                    err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                    err1 = fabs(sRGBmap(resultPtr[0])
+                                                - sRGBmap(expected[0]));
+                                    err2 = fabs(sRGBmap(resultPtr[1])
+                                                - sRGBmap(expected[1]));
+                                    err3 = fabs(sRGBmap(resultPtr[2])
+                                                - sRGBmap(expected[2]));
                                     err4 = fabsf( resultPtr[3] - expected[3] );
                                 }
                             }
@@ -674,9 +680,12 @@ int test_read_image_1D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                       xOffsetValues[ j ], yOffsetValues[ j ], 0.f, norm_offset_x, norm_offset_y, 0.0f,
                                                                                       imageSampler, expected, 0, &containsDenormals, lod );
 
-                                float err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                float err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                float err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                float err1 = fabs(sRGBmap(resultPtr[0])
+                                                  - sRGBmap(expected[0]));
+                                float err2 = fabs(sRGBmap(resultPtr[1])
+                                                  - sRGBmap(expected[1]));
+                                float err3 = fabs(sRGBmap(resultPtr[2])
+                                                  - sRGBmap(expected[2]));
                                 float err4 = fabsf( resultPtr[3] - expected[3] );
 
                                 float maxErr = 0.6;
@@ -695,9 +704,12 @@ int test_read_image_1D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                    xOffsetValues[ j ], yOffsetValues[ j ], 0.f, norm_offset_x, norm_offset_y, 0.0f,
                                                                                    imageSampler, expected, 0, NULL, lod );
 
-                                        err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                        err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                        err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                        err1 = fabs(sRGBmap(resultPtr[0])
+                                                    - sRGBmap(expected[0]));
+                                        err2 = fabs(sRGBmap(resultPtr[1])
+                                                    - sRGBmap(expected[1]));
+                                        err3 = fabs(sRGBmap(resultPtr[2])
+                                                    - sRGBmap(expected[2]));
                                         err4 = fabsf( resultPtr[3] - expected[3] );
                                     }
                                 }
