+++ test_conformance/images/kernel_read_write/test_read_2D_array.cpp
@@ -757,9 +757,12 @@ int test_read_image_2D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                           norm_offset_x, norm_offset_y, norm_offset_z,
                                                                                           imageSampler, expected, 0, &hasDenormals, lod );
 
-                                    float err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                    float err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                    float err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                    float err1 = fabs(sRGBmap(resultPtr[0])
+                                                      - sRGBmap(expected[0]));
+                                    float err2 = fabs(sRGBmap(resultPtr[1])
+                                                      - sRGBmap(expected[1]));
+                                    float err3 = fabs(sRGBmap(resultPtr[2])
+                                                      - sRGBmap(expected[2]));
                                     float err4 = fabsf( resultPtr[3] - expected[3] );
                                     float maxErr = 0.5;
 
@@ -777,9 +780,12 @@ int test_read_image_2D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                        norm_offset_x, norm_offset_y, norm_offset_z,
                                                                                        imageSampler, expected, 0, NULL, lod );
 
-                                            err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                            err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                            err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                            err1 = fabs(sRGBmap(resultPtr[0])
+                                                        - sRGBmap(expected[0]));
+                                            err2 = fabs(sRGBmap(resultPtr[1])
+                                                        - sRGBmap(expected[1]));
+                                            err3 = fabs(sRGBmap(resultPtr[2])
+                                                        - sRGBmap(expected[2]));
                                             err4 = fabsf( resultPtr[3] - expected[3] );
                                         }
                                     }
@@ -805,9 +811,15 @@ int test_read_image_2D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                               norm_offset_x, norm_offset_y, norm_offset_z,
                                                                                               imageSampler, expected, 0, &hasDenormals, lod );
 
-                                        float err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                        float err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                        float err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                        float err1 =
+                                            fabs(sRGBmap(resultPtr[0])
+                                                 - sRGBmap(expected[0]));
+                                        float err2 =
+                                            fabs(sRGBmap(resultPtr[1])
+                                                 - sRGBmap(expected[1]));
+                                        float err3 =
+                                            fabs(sRGBmap(resultPtr[2])
+                                                 - sRGBmap(expected[2]));
                                         float err4 = fabsf( resultPtr[3] - expected[3] );
                                         float maxErr = 0.6;
 
@@ -824,9 +836,15 @@ int test_read_image_2D_array( cl_context context, cl_command_queue queue, cl_ker
                                                                                     xOffsetValues[ j ], yOffsetValues[ j ], zOffsetValues[ j ],
                                                                                     imageSampler, expected, 0, NULL, lod );
 
-                                                err1 = fabsf( sRGBmap( resultPtr[0] ) - sRGBmap( expected[0] ) );
-                                                err2 = fabsf( sRGBmap( resultPtr[1] ) - sRGBmap( expected[1] ) );
-                                                err3 = fabsf( sRGBmap( resultPtr[2] ) - sRGBmap( expected[2] ) );
+                                                err1 = fabs(
+                                                    sRGBmap(resultPtr[0])
+                                                    - sRGBmap(expected[0]));
+                                                err2 = fabs(
+                                                    sRGBmap(resultPtr[1])
+                                                    - sRGBmap(expected[1]));
+                                                err3 = fabs(
+                                                    sRGBmap(resultPtr[2])
+                                                    - sRGBmap(expected[2]));
                                                 err4 = fabsf( resultPtr[3] - expected[3] );
                                             }
                                         }
